fix(migrations): utiliser le bigint historique des clients

+++ database/migrations/2026_07_17_222000_create_credit_notes_tables.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedInteger('company_id');
             $table->unsignedInteger('invoice_id');
-            $table->unsignedInteger('customer_id');
+            $table->unsignedBigInteger('customer_id');
             $table->unsignedInteger('creator_id')->nullable();
             $table->unsignedInteger('currency_id');
             $table->string('credit_note_number', 40);
